backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16.0,<4.0
PyMySQL>=1.1.1,<2.0
reportlab>=4.2.2,<5.0
gunicorn>=23.0.0,<24.0
whitenoise>=6.8.2,<7.0
